feat: add bottom sheet modal in SessionControl component
